requirements.txt
PyAudio==0.2.11
python-chess==1.0.1
SpeechRecognition==3.8.1
